src/control-plane-api/pyproject.toml
[tool.poetry]
name = "control-plane-api"
version = "0.1.0"
description = "CML Cloud Manager - Control Plane API with UI"
authors = ["Bruno van de Werve <bvandewe@gmail.com>"]
readme = "README.md"
packages = [{include = "api"}, {include = "application"}, {include = "domain"}, {include = "infrastructure"}, {include = "integration"}, {include = "observability"}, {include = "ui"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
neuroglia-python = ">=0.6.9,<0.7.0"
python-keycloak = ">=5.8.1,<6.0.0"
itsdangerous = ">=2.2.0,<3.0.0"
redis = ">=7.0.1,<8.0.0"
motor = "^3.7.1"
httpx = ">=0.24.0,<0.28.0"
boto3 = "^1.40.74"
apscheduler = "^3.11.1"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
black = ">=23.0.0"
ruff = ">=0.1.0"
mypy = "^1.11.0"
pytest-cov = "^5.0.0"
pre-commit = "^4.3.0"
detect-secrets = "^1.5.0"
pytest-asyncio = "^1.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
select = ["E", "F", "W", "I", "UP"]

[tool.black]
line-length = 120

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
    "api: marks tests as API tests",
    "e2e: marks tests as end-to-end tests",
]
